Committing first version
This is just a small tool I wrote to call DLL functions similar to
"rundll32.exe" in Windows.  It loads a DLL, locates a function, pushes
optional arguments onto the stack, invokes the function and returns the
result as the exit code.  Eventually I'd like to have a way for the
returned result to be specified as either int, float, string, etc. and
then printed out to the command line for programs that could leverage
this tool and the DLL function's results.